Move class into test file
