+++ tests/Cache/CacheArrayStoreTest.php
@@ -205,3 +205,20 @@ class CacheArrayStoreTest extends TestCase
         $this->assertTrue($unserialized->wasSerialized);
     }
 }
+
+use Serializable;
+
+class SerializableChecker implements Serializable
+{
+    public $wasSerialized = false;
+
+    public function serialize()
+    {
+        return serialize('');
+    }
+
+    public function unserialize($serialized)
+    {
+        $this->wasSerialized = true;
+    }
+}
